Fix path change with most recent node
Turns out it was not __dirname, but the way that we passed $(dirname $0)
through which ended up using /tmp (I presume that node has changed how
it launches scripts between versions). I've checked running the server
from both the server root and the test directory and the path resolves
correctly both ways around. I've printed the resolved version to the
console though to help with debugging.